backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
sentence-transformers==2.2.2
spacy==3.7.0
faiss-cpu==1.7.4
numpy==1.27.0
python-multipart==0.0.6
pydantic==1.10.12
